Revert "Cosmetic changes plus amd64 fix (hopefully)"
This reverts commit cab561c82809f7a3e1c34784d6018ecf3551fa07.

# Conflicts:
#	engine/opengl/Driver.cpp

+++ engine/sprite.cpp
@@ -8,7 +8,9 @@
 #include "video/Driver.h"
 #include "video/Image.h"
 
-Sprite::Sprite(const std::string& fname, Video::Driver* v) : video(v) {
+Sprite::Sprite(const std::string& fname, Video::Driver* v)
+    : video(v)
+{
     CCHRfile chr;
     chr.Load(fname);
         
@@ -28,7 +30,8 @@ Sprite::Sprite(const std::string& fname, Video::Driver* v) : video(v) {
         "upleft", "upright", "downleft", "downright"
     };
     const int numDirs = sizeof dirNames / sizeof dirNames[0];
-    for (uint i = 0; i < (uint)numDirs; i++) {
+    for (uint i = 0; i < numDirs; i++)
+    {
         std::string s;
         
         s = "idle_";    s += dirNames[i];   _idleScripts[i] = _scripts[s];
@@ -36,7 +39,8 @@ Sprite::Sprite(const std::string& fname, Video::Driver* v) : video(v) {
     }
     
     _frames.resize(chr.NumFrames());
-    for (uint i = 0; i < chr.NumFrames(); i++) {
+    for (uint i = 0; i < chr.NumFrames(); i++)
+    {
         _frames[i] = video->CreateImage(chr.GetFrame(i));
     }
 }
